+++ src/actions/BackupModalActions.tsx
@@ -2,14 +2,10 @@ import { EdgeAccount } from 'edge-core-js'
 import * as React from 'react'
 import { AirshipBridge } from 'react-native-airship'
 
-import { BackupForTransferModal, BackupForTransferModalResult } from '../components/modals/BackupForTransferModal'
-import { BackupModal, BackupModalResult } from '../components/modals/BackupModal'
+import { BackupForAccountModal, BackupForTransferModal, BackupForTransferModalResult, BackupModalResult } from '../components/modals/BackupModal'
 import { Airship, showError } from '../components/services/AirshipInstance'
 import { getExperimentConfigValue } from '../experimentConfig'
 import { NavigationBase } from '../types/routerTypes'
-
-// TODO: Merge our various backup modal flavors with a common design that
-// satisfies all requirements (design TBD).
 
 let isBackupModalShowing = false
 
@@ -24,7 +20,7 @@ export const showBackupModal = (props: { navigation: NavigationBase; forgetLogin
   const { navigation, forgetLoginId } = props
 
   Airship.show((bridge: AirshipBridge<BackupModalResult | undefined>) => {
-    return <BackupModal bridge={bridge} forgetLoginId={forgetLoginId} />
+    return <BackupForAccountModal bridge={bridge} forgetLoginId={forgetLoginId} />
   })
     .then((userSel?: BackupModalResult) => {
       if (userSel === 'upgrade') {
