+++ src/components/scenes/DevTestScene.tsx
@@ -18,7 +18,7 @@ import { consify } from '../../util/utils'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { styled } from '../hoc/styled'
 import { SearchIconAnimated } from '../icons/ThemedIcons'
-import { BackupForTransferModal, BackupForTransferModalResult } from '../modals/BackupForTransferModal'
+import { BackupForTransferModal, BackupForTransferModalResult } from '../modals/BackupModal'
 import { ButtonsModal } from '../modals/ButtonsModal'
 import { ConfirmContinueModal } from '../modals/ConfirmContinueModal'
 import { CountryListModal } from '../modals/CountryListModal'
